requirements.txt
Flask
requests
python-doniclient
python-openstackclient
keystoneauth1